requirements.txt
streamlit
pandas
matplotlib
seaborn
wordcloud
emoji
nltk
google-generativeai
python-dotenv
